+++ rust/src/lib.rs
@@ -97,8 +97,8 @@ pub mod partitions;
 pub mod schema;
 pub mod storage;
 mod table_state;
-pub mod write;
 pub mod writer;
+mod time_utils;
 
 #[cfg(feature = "datafusion-ext")]
 pub mod delta_datafusion;
