+++ rust/src/writer/test_utils.rs
